==> JupyterLab/LAM/3_counMatrixGen_COMPARE.ipynb <==
#SETTING PATH 
path="/sharedFolder/Data/1_HTGTS/13_Revision1/pooled/"
previousResults="/sharedFolder/Results/Hotspot/Revision1/FULL/"
results="/sharedFolder/Results/Hotspot/Revision1/FULL/"
dir.create(results,recursive=TRUE)
riderPath="/home/Riderv2/"

#LOAD USEFULL FUNCTIONS 
noExt=function(x){
return(tools::file_path_sans_ext(x))
}
bnoExt=function(x){
return(basename(tools::file_path_sans_ext(x)))
}
fNumber=function(x){
sapply(x,FUN=function(x){
strsplit(x,"__")[[1]][2]
})
}
rRider=function(x){
sapply(x,FUN=function(x){
strsplit(x,"-RIDER")[[1]][1]
})
}
# centering with 'scale()'
center_scale <- function(x) {
    scale(x, scale = FALSE)
}

#LOAD LIBRARIES 
library("argparser");
library(gtools);
library("GenomicRanges");
library("ggVennDiagram");
options(warn=-1)
library(GenomicRanges)

print("hey")
#Load Data
for(jamesonFirst in c(5,6,7)){
Database=read.table(paste("/sharedFolder/Results/Hotspot/Revision1/FULL/",jamesonFirst,"/Database_annotated.csv",sep=""),header=TRUE,sep=",")
a=list.files(path,pattern=".bed",recursive=TRUE)
a=mixedsort(a)
a2=sapply(a,FUN=function(x){strsplit(x,".bed")[[1]][1]})
a2=sapply(a2,FUN=function(x){strsplit(x,"__")[[1]][2]})
a=list.files(path,pattern=".bed",recursive=TRUE)
a=mixedsort(a)
a2=sapply(a,FUN=function(x){strsplit(x,".bed")[[1]][1]})
a2=sapply(a2,FUN=function(x){strsplit(x,"__")[[1]][2]})
hgtgts_count=matrix(0,ncol=length(unique(Database[,7])),nrow=length(a2))
colnames(hgtgts_count)=sapply(unique(Database[,7]),FUN=toString)
rownames(hgtgts_count)=a2
    
    
hgtgts_rpm=hgtgts_count
hgtgts_fpkm=hgtgts_count
for(j in a){
    system(paste("echo",j))
temp=read.table(paste(path,"/",j,sep=""),sep="\t")
tempGG=GRanges(seqnames=as.matrix(temp[,1]),ranges=IRanges(start = as.numeric(temp[,2]),end=as.numeric(temp[,3])))    
    for(i in seq(nrow(Database))){
        system(paste("echo",i/nrow(Database)))
        gg=GRanges(seqnames=as.matrix(Database[i,1]),ranges=IRanges(start = as.numeric(Database[i,2]),end=as.numeric(Database[i,3])))    
        over=countOverlaps(gg,tempGG)
        name=as.vector(noExt(fNumber(j)))
        hgtgts_count[name,toString(Database[i,7])]= over
        hgtgts_rpm[name,toString(Database[i,7])]= ((over)/nrow(temp))*1000000
   }

}
dir.create(paste(results,"/",jamesonFirst,"/RPMCOmparison/",sep=""))

    
    
    

write.table((t(hgtgts_count)),paste(results,"/",jamesonFirst,"/RPMCOmparison/","FULLcount_hgtgts.csv",sep=""),col.names=NA,sep=",")
write.table((t(hgtgts_rpm)),paste(results,"/",jamesonFirst,"/RPMCOmparison/","FULLrpm_hgtgts.csv",sep=""),col.names=NA,sep=",")

        
    
annotations=c()    
 for(n in colnames(hgtgts_count)){
     ann=unique(Database[which(Database$hotspotID==n),]$annotation)
     if(ann=="NULL"){
     ann=paste(unique(Database[which(Database$hotspotID==n),]$Chromosome),":",unique(Database[which(Database$hotspotID==n),]$ChrStart),"-",unique(Database[which(Database$hotspotID==n),]$ChrEnd),sep="")
     }
     annotations=c(annotations,ann)
 }   
    
colnames(hgtgts_count)=annotations
colnames(hgtgts_rpm)=annotations
    
write.table((t(hgtgts_count)),paste(results,"/",jamesonFirst,"/RPMCOmparison/","FULLcount_hgtgtsANNOTATED.csv",sep=""),col.names=NA,sep=",")
write.table((t(hgtgts_rpm)),paste(results,"/",jamesonFirst,"/RPMCOmparison/","FULLrpm_hgtgtsANNOTATED.csv",sep=""),col.names=NA,sep=",")    
}

print("hey")
#Load Data
for(jamesonFirst in c(5,6,7)){
Database=read.table(paste("/sharedFolder/Results/Hotspot/Revision1/FULL/",jamesonFirst,"/Database_annotated.csv",sep=""),header=TRUE,sep=",")
Database=Database[which(Database$Category=="AID_off_target"),]
a=list.files(path,pattern=".bed",recursive=TRUE)
a=mixedsort(a)
a2=sapply(a,FUN=function(x){strsplit(x,".bed")[[1]][1]})
a2=sapply(a2,FUN=function(x){strsplit(x,"__")[[1]][2]})
a=list.files(path,pattern=".bed",recursive=TRUE)
a=mixedsort(a)
a2=sapply(a,FUN=function(x){strsplit(x,".bed")[[1]][1]})
a2=sapply(a2,FUN=function(x){strsplit(x,"__")[[1]][2]})
hgtgts_count=matrix(0,ncol=length(unique(Database[,7])),nrow=length(a2))
colnames(hgtgts_count)=sapply(unique(Database[,7]),FUN=toString)
rownames(hgtgts_count)=a2
    
    
hgtgts_rpm=hgtgts_count
hgtgts_fpkm=hgtgts_count
for(j in a){
    system(paste("echo",j))
temp=read.table(paste(path,"/",j,sep=""),sep="\t")
tempGG=GRanges(seqnames=as.matrix(temp[,1]),ranges=IRanges(start = as.numeric(temp[,2]),end=as.numeric(temp[,3])))    
    for(i in seq(nrow(Database))){
        system(paste("echo",i/nrow(Database)))
        gg=GRanges(seqnames=as.matrix(Database[i,1]),ranges=IRanges(start = as.numeric(Database[i,2]),end=as.numeric(Database[i,3])))    
        over=countOverlaps(gg,tempGG)
        name=as.vector(noExt(fNumber(j)))
        hgtgts_count[name,toString(Database[i,7])]= over
        hgtgts_rpm[name,toString(Database[i,7])]= ((over)/nrow(temp))*1000000
   }

}
dir.create(paste(results,"/",jamesonFirst,"/RPMCOmparison/",sep=""))

    
    
    

write.table((t(hgtgts_count)),paste(results,"/",jamesonFirst,"/RPMCOmparison/","count_hgtgts.csv",sep=""),col.names=NA,sep=",")
write.table((t(hgtgts_rpm)),paste(results,"/",jamesonFirst,"/RPMCOmparison/","rpm_hgtgts.csv",sep=""),col.names=NA,sep=",")

      
  hgtgts_rpmDMSO=hgtgts_rpm[,sort((hgtgts_rpm["MEC-1_Idelalisib_DMSO",]+0.00000001)/(hgtgts_rpm["AID-WT_DMSO",]+0.00000001),index.return=TRUE,decreasing=TRUE)$ix]
  write.table((t(hgtgts_rpmDMSO)),paste(results,"/",jamesonFirst,"/RPMCOmparison/","rpm_hgtgtsHeatmapDMSO.csv",sep=""),col.names=NA,sep=",")      
    
annotations=c()    
 for(n in colnames(hgtgts_count)){
     ann=unique(Database[which(Database$hotspotID==n),]$annotation)
     if(ann=="NULL"){
     ann=paste(unique(Database[which(Database$hotspotID==n),]$Chromosome),":",unique(Database[which(Database$hotspotID==n),]$ChrStart),"-",unique(Database[which(Database$hotspotID==n),]$ChrEnd),sep="")
     }
     annotations=c(annotations,ann)
 }   
    
colnames(hgtgts_count)=annotations
colnames(hgtgts_rpm)=annotations
    
write.table((t(hgtgts_count)),paste(results,"/",jamesonFirst,"/RPMCOmparison/","count_hgtgtsANNOTATED.csv",sep=""),col.names=NA,sep=",")
write.table((t(hgtgts_rpm)),paste(results,"/",jamesonFirst,"/RPMCOmparison/","rpm_hgtgtsANNOTATED.csv",sep=""),col.names=NA,sep=",")  
    

    
}

library(beeswarm)

thresholds <- c(5,6,7)

for (jamesonFirst in thresholds) {
  
  message("Processing threshold: ", jamesonFirst)
  
  data_file <- paste(results, "/", jamesonFirst, "/RPMCOmparison/", "count_hgtgtsANNOTATED.csv", sep = "")
  data <- read.csv(data_file, header = TRUE, row.names = 1, sep = ",")
  data <- data + 1  # Pseudocount
  
  # Calcolo Fold Change
  foldChange_IDE_DMSO <- data[,"MEC.1_Idelalisib_DMSO"] / data[,"AID.WT_DMSO"]
  foldChange_IDE_Taze <- data[,"MEC.1_Idelalisib_Tazemetostat"] / data[,"AID.WT_DMSO"]
  foldChange_IDE_Vale <- data[,"MEC.1_Idelalisib_Valemetostat"] / data[,"AID.WT_DMSO"]
  foldChange_AID_WT_Taze <- data[,"AID.WT_Tazemetostat"] / data[,"AID.WT_DMSO"]
  foldChange_AID_WT_Vale <- data[,"AID.WT_Valemetostat"] / data[,"AID.WT_DMSO"]
  dmso_WT <- data[,"AID.WT_DMSO"] / data[,"AID.WT_DMSO"]
  
  foldChange_table <- data.frame(
    DMSO = dmso_WT,
    Tazemetostat = foldChange_AID_WT_Taze,
    Valemetostat = foldChange_AID_WT_Vale,
    Idelalisib = foldChange_IDE_DMSO,
    `Idelalisib+Tazemetostat` = foldChange_IDE_Taze,
    `Idelalisib+Valemetostat` = foldChange_IDE_Vale
  )
  
  # Output dei dati fold change
  out_table <- paste(results, "/", jamesonFirst, "/RPMCOmparison/", "boxPlotCount_", jamesonFirst, "way.csv", sep = "")
  write.table(foldChange_table, out_table, col.names = TRUE, row.names = FALSE, sep = ",")
  
  # Plot
  #png(filename = paste(results, "/", jamesonFirst, "/RPMCOmparison/", "foldchange_boxplotCount_", jamesonFirst, ".png", sep = ""), width = 800, height = 600)
  
   png(
  filename = paste(results, "/", jamesonFirst, "/RPMCOmparison/", "foldchange_boxplotCount_", jamesonFirst, ".png", sep = ""),
  width = 2000,
  height = 1500,
  res = 300,
  type = "cairo"
)
 
    
    
  boxplot(foldChange_table, xaxt = "n", main = paste("Fold Change - threshold", jamesonFirst),
          xlab = "", ylab = "Fold Change", outline = FALSE, col = "gray90", border = "gray40")
  


    
    
  for (i in 1:ncol(foldChange_table)) {
    color <- if (grepl("Idelalisib", colnames(foldChange_table)[i], ignore.case = TRUE)) "red" else "blue"
    beeswarm(foldChange_table[[i]], add = TRUE, at = i, col = color, pch = 20, vertical = TRUE)
  }
  
  labels <- colnames(foldChange_table)
  axis(1, at = 1:length(labels), labels = FALSE)
  text(x = 1:length(labels), y = par("usr")[3] - 0.3, srt = 45, adj = 1,
       labels = labels, xpd = TRUE, cex = 0.7)
  
  dev.off()
}


thresholds <- c(5, 6, 7)

for (jamesonFirst in thresholds) {
  
  message("Processing threshold: ", jamesonFirst)
  
  data_file <- paste(results, "/", jamesonFirst, "/RPMCOmparison/", "count_hgtgtsANNOTATED.csv", sep = "")
  a <- read.csv(data_file, header = TRUE, row.names = 1, sep = ",")
  
  conditions <- colnames(a)
  reference_condition <- a$AID.WT_DMSO
  
  # Inizializza vettore per p-value
  p_values <- numeric(length(conditions))
  names(p_values) <- conditions
  
  # Test Wilcoxon vs AID.WT_DMSO
  for (i in 1:length(conditions)) {
    cond <- conditions[i]
    if (cond != "AID.WT_DMSO") {
      p_values[cond] <- wilcox.test(a[[cond]], reference_condition)$p.value
    } else {
      p_values[cond] <- NA  # Reference vs se stessa
    }
  }

  # Confronti Idelalisib combinato vs Idelalisib DMSO
  p_values[["Idelalisib_Taze_vs_Idelalisib_DMSO"]] <- wilcox.test(a$MEC.1_Idelalisib_Tazemetostat, a$MEC.1_Idelalisib_DMSO)$p.value
  p_values[["Idelalisib_Vale_vs_Idelalisib_DMSO"]] <- wilcox.test(a$MEC.1_Idelalisib_Valemetostat, a$MEC.1_Idelalisib_DMSO)$p.value

  # Confronti Idelalisib combinato vs WT combinato
  p_values[["Idelalisib_Taze_vs_WT_Taze"]] <- wilcox.test(a$MEC.1_Idelalisib_Tazemetostat, a$AID.WT_Tazemetostat)$p.value
  p_values[["Idelalisib_Vale_vs_WT_Vale"]] <- wilcox.test(a$MEC.1_Idelalisib_Valemetostat, a$AID.WT_Valemetostat)$p.value

  # Stampa
  cat("\nP-values for threshold", jamesonFirst, ":\n")
  for (n in names(p_values)) {
    cat(n, ":", p_values[n], "\n")
  }

  # Salvataggio su file
  pval_out <- paste(results, "/", jamesonFirst, "/RPMCOmparison/", "pvaluesCOUNT_", jamesonFirst, ".csv", sep = "")
  write.csv(as.data.frame(p_values), file = pval_out, quote = FALSE)
}


library(beeswarm)

thresholds <- c(5,6,7)

for (jamesonFirst in thresholds) {
  
  message("Processing threshold: ", jamesonFirst)
  
  data_file <- paste(results, "/", jamesonFirst, "/RPMCOmparison/", "rpm_hgtgtsANNOTATED.csv", sep = "")
  data <- read.csv(data_file, header = TRUE, row.names = 1, sep = ",")
  data <- data + 1  # Pseudocount
  
  # Calcolo Fold Change
  foldChange_IDE_DMSO <- data[,"MEC.1_Idelalisib_DMSO"] / data[,"AID.WT_DMSO"]
  foldChange_IDE_Taze <- data[,"MEC.1_Idelalisib_Tazemetostat"] / data[,"AID.WT_DMSO"]
  foldChange_IDE_Vale <- data[,"MEC.1_Idelalisib_Valemetostat"] / data[,"AID.WT_DMSO"]
  foldChange_AID_WT_Taze <- data[,"AID.WT_Tazemetostat"] / data[,"AID.WT_DMSO"]
  foldChange_AID_WT_Vale <- data[,"AID.WT_Valemetostat"] / data[,"AID.WT_DMSO"]
  dmso_WT <- data[,"AID.WT_DMSO"] / data[,"AID.WT_DMSO"]
  
  foldChange_table <- data.frame(
    DMSO = dmso_WT,
    Tazemetostat = foldChange_AID_WT_Taze,
    Valemetostat = foldChange_AID_WT_Vale,
    Idelalisib = foldChange_IDE_DMSO,
    `Idelalisib+Tazemetostat` = foldChange_IDE_Taze,
    `Idelalisib+Valemetostat` = foldChange_IDE_Vale
  )
  
  # Output dei dati fold change
  out_table <- paste(results, "/", jamesonFirst, "/RPMCOmparison/", "boxPlotRPM_", jamesonFirst, "way.csv", sep = "")
  write.table(foldChange_table, out_table, col.names = TRUE, row.names = FALSE, sep = ",")
  
  # Plot
  #png(filename = paste(results, "/", jamesonFirst, "/RPMCOmparison/", "foldchange_boxplot_RPM", jamesonFirst, ".png", sep = ""), width = 800, height = 600)
png(
  filename = paste(results, "/", jamesonFirst, "/RPMCOmparison/", "foldchange_boxplot_RPM", jamesonFirst, ".png", sep = ""),
  width = 2000,
  height = 1500,
  res = 300,
  type = "cairo"
)


  boxplot(foldChange_table, xaxt = "n", main = paste("Fold Change - threshold", jamesonFirst),
          xlab = "", ylab = "Fold Change", outline = FALSE, col = "gray90", border = "gray40")
  
  for (i in 1:ncol(foldChange_table)) {
    color <- if (grepl("Idelalisib", colnames(foldChange_table)[i], ignore.case = TRUE)) "red" else "blue"
    beeswarm(foldChange_table[[i]], add = TRUE, at = i, col = color, pch = 20, vertical = TRUE)
  }
  
  labels <- colnames(foldChange_table)
  axis(1, at = 1:length(labels), labels = FALSE)
  text(x = 1:length(labels), y = par("usr")[3] - 0.3, srt = 45, adj = 1,
       labels = labels, xpd = TRUE, cex = 0.7)
  
  dev.off()
}


thresholds <- c(5, 6, 7)

for (jamesonFirst in thresholds) {
  
  message("Processing threshold: ", jamesonFirst)
  
  data_file <- paste(results, "/", jamesonFirst, "/RPMCOmparison/", "rpm_hgtgtsANNOTATED.csv", sep = "")
  a <- read.csv(data_file, header = TRUE, row.names = 1, sep = ",")
  
  conditions <- colnames(a)
  reference_condition <- a$AID.WT_DMSO
  
  # Inizializza vettore per p-value
  p_values <- numeric(length(conditions))
  names(p_values) <- conditions
  
  # Test Wilcoxon vs AID.WT_DMSO
  for (i in 1:length(conditions)) {
    cond <- conditions[i]
    if (cond != "AID.WT_DMSO") {
      p_values[cond] <- wilcox.test(a[[cond]], reference_condition)$p.value
    } else {
      p_values[cond] <- NA  # Reference vs se stessa
    }
  }

  # Confronti Idelalisib combinato vs Idelalisib DMSO
  p_values[["Idelalisib_Taze_vs_Idelalisib_DMSO"]] <- wilcox.test(a$MEC.1_Idelalisib_Tazemetostat, a$MEC.1_Idelalisib_DMSO)$p.value
  p_values[["Idelalisib_Vale_vs_Idelalisib_DMSO"]] <- wilcox.test(a$MEC.1_Idelalisib_Valemetostat, a$MEC.1_Idelalisib_DMSO)$p.value

  # Confronti Idelalisib combinato vs WT combinato
  p_values[["Idelalisib_Taze_vs_WT_Taze"]] <- wilcox.test(a$MEC.1_Idelalisib_Tazemetostat, a$AID.WT_Tazemetostat)$p.value
  p_values[["Idelalisib_Vale_vs_WT_Vale"]] <- wilcox.test(a$MEC.1_Idelalisib_Valemetostat, a$AID.WT_Valemetostat)$p.value

  # Stampa
  cat("\nP-values for threshold", jamesonFirst, ":\n")
  for (n in names(p_values)) {
    cat(n, ":", p_values[n], "\n")
  }

  # Salvataggio su file
  pval_out <- paste(results, "/", jamesonFirst, "/RPMCOmparison/", "pvaluesRPM_", jamesonFirst, ".csv", sep = "")
  write.csv(as.data.frame(p_values), file = pval_out, quote = FALSE)
}
