==> JupyterLab/LAM/5_FAIR.ipynb <==
#SETTING PATH 

for(jamesonFirst in c(5,6,7)){
path="/sharedFolder/Data/1_HTGTS/13_Revision1/pooled/"
previousResults=paste("/sharedFolder/Results/Hotspot/Revision1/FULL/",jamesonFirst,sep="")
results=paste("/sharedFolder/Results/Hotspot/Revision1/FULL/",jamesonFirst,sep="")
dir.create(results,recursive=TRUE)
riderPath="/home/Riderv2/"
#LOAD USEFULL FUNCTIONS 
noExt=function(x){
return(tools::file_path_sans_ext(x))
}
bnoExt=function(x){
return(basename(tools::file_path_sans_ext(x)))
}
fNumber=function(x){
sapply(x,FUN=function(x){
strsplit(x,"__")[[1]][2]
})
}
rRider=function(x){
sapply(x,FUN=function(x){
strsplit(x,"-RIDER")[[1]][1]
})
}
# centering with 'scale()'
center_scale <- function(x) {
    scale(x, scale = FALSE)
}
#LOAD LIBRARIES 
library("argparser");
library(gtools);
library("GenomicRanges");
library("ggVennDiagram");
library("ggplot2")
library(eulerr)
options(warn=-1)
library(GenomicRanges)
#Load Data
Database=read.table(paste(previousResults,"Database.csv",sep="/"),header=TRUE,sep=",")
Database=Database[which(Database$Category=="AID_off_target"),]

a=list.files(path,pattern=".bed",recursive=TRUE)
a=mixedsort(a)
a2=sapply(a,FUN=function(x){strsplit(x,".bed")[[1]][1]})
a2=sapply(a2,FUN=function(x){strsplit(x,"__")[[1]][2]})
conditions=c("DMSO","Taze","Vale")
tempWT=Database[grep("WT",Database[,6]),]
tempIde=Database[grep("Idelalisib",Database[,6]),]
ll=list()
lll=list()
count=0
for(i in conditions){
count=count+1
wt=tempWT[grep(i,tempWT[,6]),]
id=tempIde[grep(i,tempIde[,6]),]
wt1=wt[,7]
names(wt1)=wt[,6]
id1=id[,7]
names(id1)=id[,6]    
fit1 <- euler(c("WT" = length(setdiff(wt1,id1)), "Ide" = length(setdiff(id1,wt1)), "WT&Ide" = length(intersect(wt1,id1))))
ll[[count]]=fit1
lll[[count]]=unique(union(wt1,id1))
}
names(lll)=conditions
heatmap=read.table(paste(results,"/RPMCOmparison/rpm_hgtgtsHeatmapDMSO.csv",sep=""),header=TRUE,row.names=1,sep=",")

for(i in conditions){
    tmp=heatmap[,grep(i,colnames(heatmap))]
    tmp=tmp[as.character(lll[[i]]),]
    tmp=tmp[sort(tmp[,3],index.return=TRUE,decreasing=TRUE)$ix,]
    write.table(tmp, paste(results,"/RPMCOmparison/",i,"rpm_FAIR.csv",sep=""),sep=",",col.names=NA)
}

    
    
    data1=read.table(paste(results,"/RPMCOmparison/","DMSOrpm_FAIR.csv",sep=""),header=TRUE,sep=",",row.names=1)
data2=read.table(paste(previousResults,"Database_annotated.csv",sep="/"),header=TRUE,sep=",")
newNames=c()
for(i in rownames(data1)){
    if(unique(data2[which(data2$hotspotID==i),"annotation"])=="NULL"){newNames=append(newNames,paste(unique(data2[which(data2$hotspotID==i),c("Chromosome","ChrStart","ChrEnd")]),collapse="_"))}else if(length(strsplit(unique(data2[which(data2$hotspotID==i),"annotation"]),"_")[[1]])>3){ 
        newNames=append(newNames,paste(unique(data2[which(data2$hotspotID==i),c("Chromosome","ChrStart","ChrEnd")]),collapse="_"))
    }
    else{
newNames=append(newNames,unique(data2[which(data2$hotspotID==i),"annotation"]))}
}
rownames(data1)=newNames
increased=which(data1$MEC.1_Idelalisib_DMSO - data1$AID.WT_DMSO>0)
print(length(increased)/nrow(data1))
print(paste(length(increased),"increased over",nrow(data1)))
write.table(data1[increased,],paste(results,"/RPMCOmparison/","DMSOrpm_FAIR_ANNOTATED.csv",sep=""),col.names=NA,sep=",")

    data1=read.table(paste(results,"/RPMCOmparison/","Tazerpm_FAIR.csv",sep=""),header=TRUE,sep=",",row.names=1)
data2=read.table(paste(previousResults,"Database_annotated.csv",sep="/"),header=TRUE,sep=",")
newNames=c()
for(i in rownames(data1)){
    if(unique(data2[which(data2$hotspotID==i),"annotation"])=="NULL"){newNames=append(newNames,paste(unique(data2[which(data2$hotspotID==i),c("Chromosome","ChrStart","ChrEnd")]),collapse="_"))}else if(length(strsplit(unique(data2[which(data2$hotspotID==i),"annotation"]),"_")[[1]])>3){ 
        newNames=append(newNames,paste(unique(data2[which(data2$hotspotID==i),c("Chromosome","ChrStart","ChrEnd")]),collapse="_"))
    }
    else{
newNames=append(newNames,unique(data2[which(data2$hotspotID==i),"annotation"]))}
}
rownames(data1)=newNames
increased=which(data1$MEC.1_Idelalisib_Tazemetostat - data1$AID.WT_Tazemetostat	>0)
print(length(increased)/nrow(data1))
print(paste(length(increased),"increased over",nrow(data1)))
write.table(data1[increased,],paste(results,"/RPMCOmparison/","Tazerpm_FAIR_ANNOTATED.csv",sep=""),col.names=NA,sep=",")

    data1=read.table(paste(results,"/RPMCOmparison/","Valerpm_FAIR.csv",sep=""),header=TRUE,sep=",",row.names=1)
data2=read.table(paste(previousResults,"Database_annotated.csv",sep="/"),header=TRUE,sep=",")
newNames=c()
for(i in rownames(data1)){
    if(unique(data2[which(data2$hotspotID==i),"annotation"])=="NULL"){newNames=append(newNames,paste(unique(data2[which(data2$hotspotID==i),c("Chromosome","ChrStart","ChrEnd")]),collapse="_"))}else if(length(strsplit(unique(data2[which(data2$hotspotID==i),"annotation"]),"_")[[1]])>3){ 
        newNames=append(newNames,paste(unique(data2[which(data2$hotspotID==i),c("Chromosome","ChrStart","ChrEnd")]),collapse="_"))
    }
    else{
newNames=append(newNames,unique(data2[which(data2$hotspotID==i),"annotation"]))}
}
rownames(data1)=newNames
increased=which(data1$MEC.1_Idelalisib_Valemetostat - data1$AID.WT_Valemetostat>0)
print(length(increased)/nrow(data1))
print(paste(length(increased),"increased over",nrow(data1)))
write.table(data1[increased,],paste(results,"/RPMCOmparison/","Valerpm_FAIR_ANNOTATED.csv",sep=""),col.names=NA,sep=",")

    
}
