==> JupyterLab/LAM/6_circoPlot.ipynb <==
library(GenomicRanges)
library(data.table)
path="/sharedFolder/Data/1_HTGTS/13_Revision1/pooled/"
previousResults=paste("/sharedFolder/Results/Hotspot/Revision1/FULL/","5",sep="")
results=paste("/sharedFolder/Results/Hotspot/Revision1/FULL/","5",sep="")



gtf="/sharedFolder/Data/gene_coordinatesHG38.tsv"
gtfT=read.table(gtf,header=TRUE,sep="\t")
gtfG=GRanges(seqnames=as.matrix(gtfT[,1]),ranges=IRanges(start = as.numeric(gtfT[,2])-100,end=as.numeric(gtfT[,3])+100))


#FUNCTIONs
modify_vector <- function(vector) {
  modified_vector <- vector
  for (i in seq_along(vector)) {
    if (nchar(vector[i]) >= 10) {
      elements <- strsplit(vector[i], "_")[[1]]
      shortest_index <- which.min(nchar(elements))
      modified_vector[i] <- elements[shortest_index]
    }
  }
  return(modified_vector)
}

circoPlot=function(tempPath,increasedOnly,control,treat,threshold){
temp=read.table(tempPath,header=TRUE,row.names=1,sep=",")

#if(increasedOnly){
#temp=temp[which(temp[treat]-temp[control]>0),]
#}    
data=read.table(paste("/sharedFolder/Results/Hotspot/Revision1/FULL/",threshold,"/Database_annotated.csv",sep=""),header=TRUE,sep=",")
data=data[which(data$Category=="AID_off_target"),]

dataList=c()
for(i in rownames(temp)){
if(!is.na(grep(i,data[,9])[1])){
    dataList=append(dataList,grep(i,data[,9])[1])
    }else{
    yo=strsplit(i,"_")[[1]]
    dataList = dataList=append(dataList,which(
  data$Chromosome == yo[1] &
  data$ChrStart == as.integer(yo[2]) &
  data$ChrEnd == as.integer(yo[3])
)[1])
    }
}
    
data2=data[dataList,]
tt=c("chr8","126735434","128742951","MYC")
names(tt)=names(data2[,c(1,2,3,8)])
label=rbind(data2[,c(1,2,3,8)],t(as.matrix(tt)))
condition=c("DMSO","Taze","Vale")
forever=matrix(ncol=7)
colnames(forever)=c("chr1","start1","end1","chr2","start2","end2","score")
label[,1]=gsub("chr","hs",label[,1])
nam=c(rownames(temp),"MYC")
nam[grep("chr",nam)]="NA"
label[,4]=modify_vector(nam)

write.table(label,paste("/sharedFolder/Circos_conf/_label.txt",sep=""),col.names=FALSE,row.names=FALSE,sep="\t",quote=FALSE)
test=matrix(ncol=7)
colnames(test)=c("chr1","start1","end1","chr2","start2","end2","score")
extra_info <- cbind(rep("chr8", nrow(data2)), rep("126735434", nrow(data2)), rep("128742951", nrow(data2)))
tt=cbind(extra_info,data2[,c(1,2,3)],paste("score=",(temp[,treat]+1)/(temp[,control]+1),sep=""))
    colnames(tt)=c("chr1","start1","end1","chr2","start2","end2","score")
    forever=rbind(forever,tt)
    test=rbind(forever,tt)
    test=test[-1,]
    
    test[,1]=gsub("chr","hs",test[,1])
    test[,4]=gsub("chr","hs",test[,4])
    write.table(test,paste("/sharedFolder/Circos_conf/_link.txt",sep=""),col.names=FALSE,row.names=FALSE,sep="\t",quote=FALSE)


forever=forever[-1,]

# Sostituisci "chr" con "hs" nelle colonne "chr1" e "chr2"

forever[,1]=gsub("chr","hs",forever[,1])
forever[,4]=gsub("chr","hs",forever[,4])

write.table(forever,paste("/sharedFolder/Circos_conf/_link.txt",sep=""),col.names=FALSE,row.names=FALSE,sep="\t",quote=FALSE)
setwd(results)
system("/home/circos-0.69-9/bin/circos -conf /sharedFolder/Circos_conf/yo.conf")
print((paste("mv ",getwd(),"/circoPlot.png ",tools::file_path_sans_ext(tempPath),".png",sep="")))
system((paste("mv ",getwd(),"/circoPlot.png ",tools::file_path_sans_ext(tempPath),".png",sep=""))) 
system((paste("mv ",getwd(),"/circoPlot.svg ",tools::file_path_sans_ext(tempPath),".svg",sep=""))) 


    #system("/home/circos-0.69-9/bin/circos -conf /sharedFolder/Circos_conf/yo_NOTHICK.conf")
#system(paste("mv /sharedFolder/Results/Ligase4/Ligase4.png ",tools::file_path_sans_ext(tempPath),"_noThick.png",sep="")) 
#system(paste("mv /sharedFolder/Results/Ligase4/Ligase4.svg ",tools::file_path_sans_ext(tempPath),"_noThick.svg",sep=""))   
}

circoPlot("/sharedFolder/Results/Hotspot/Revision1/FULL/5/RPMCOmparison/DMSOrpm_FAIR_ANNOTATED.csv",FALSE,1,1,5)
circoPlot("/sharedFolder/Results/Hotspot/Revision1/FULL/5/RPMCOmparison/Tazerpm_FAIR_ANNOTATED.csv",FALSE,1,3,5)
circoPlot("/sharedFolder/Results/Hotspot/Revision1/FULL/5/RPMCOmparison/Valerpm_FAIR_ANNOTATED.csv",FALSE,1,3,5)

circoPlot("/sharedFolder/Results/Hotspot/Revision1/FULL/6/RPMCOmparison/DMSOrpm_FAIR_ANNOTATED.csv",FALSE,1,3,6)
circoPlot("/sharedFolder/Results/Hotspot/Revision1/FULL/6/RPMCOmparison/Tazerpm_FAIR_ANNOTATED.csv",FALSE,1,3,6)
circoPlot("/sharedFolder/Results/Hotspot/Revision1/FULL/6/RPMCOmparison/Valerpm_FAIR_ANNOTATED.csv",FALSE,1,3,6)

circoPlot("/sharedFolder/Results/Hotspot/Revision1/FULL/7/RPMCOmparison/DMSOrpm_FAIR_ANNOTATED.csv",FALSE,1,3,7)
circoPlot("/sharedFolder/Results/Hotspot/Revision1/FULL/7/RPMCOmparison/Tazerpm_FAIR_ANNOTATED.csv",FALSE,1,3,7)
circoPlot("/sharedFolder/Results/Hotspot/Revision1/FULL/7/RPMCOmparison/Valerpm_FAIR_ANNOTATED.csv",FALSE,1,3,7)



getwd()
